+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-backend
 
+## 2.6.37
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.37
+  - @stackframe/stack-emails@2.6.37
+
+## 2.6.36
+
+### Patch Changes
+
+- Various updates
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.36
+  - @stackframe/stack-emails@2.6.36
+
 ## 2.6.35
 
 ### Patch Changes