+++ examples/supabase/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/example-supabase
 
+## 2.6.37
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack@2.6.37
+
+## 2.6.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.36
+
 ## 2.6.35
 
 ### Patch Changes