+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/stack-emails
 
+## 2.6.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.37
+  - @stackframe/stack-ui@2.6.37
+
+## 2.6.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.36
+  - @stackframe/stack-ui@2.6.36
+
 ## 2.6.35
 
 ### Patch Changes